Massive "Vendor Sell" price tweak.
Durable and exceptional items should no longer be overvalued.
Magic items are now priced by a formula that accounts for
every individual attribute they possess (usually bonuses,
but negative attributes are computed too). The pricing is
according to utility. This is somewhat subjective, but in
this case I feel it's better to have a human touch than
none at all.