New Command: [get SellPrice
This commit introduces a GameMaster-level command:
[get SellPrice

which will provide a formula of the form
SellPrice = (BaseSellPrice * <multiplier>) + <fixed price increase>

The <multiplier> and <fixed price increase> will be numbers
computed by the item's special attributes and value as perceived
by the system. The final sell price can only be determined by
attempting to sell the item to a vendor, because the vendor's
state is part of the BaseSellPrice calculation.