Fix StackOverflowException when buying from vendor.
This happened because of infinite recursion that was triggered by
failing to update a BaseVendor's CashOnHand (as reported by
GetCashOnHand) when newly added ledger entries have equivalent
timestamps to already existing entries (simultaneity).
Simultaneity can happen when an entry happens that violates some
invariant (ex: cash-on-hand never exceeds max-cash-on-hand)
and, to satisfy that invariant, the same calculation needs to
insert an additional ledger entry at the same point in time.